dashes are not allowed in python package names, rename
