+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Install Poetry
         uses: snok/install-poetry@v1.3.4
       - name: Install Python dependencies
-        working-directory: flask-app
+        working-directory: flask_app
         run: |
           python -m pip install --upgrade pip
           poetry install
@@ -32,5 +32,5 @@ jobs:
         run: |
           npm install
       - name: Run pre-commit
-        working-directory: flask-app
+        working-directory: flask_app
         run: poetry run pre-commit run --all-files
