+++ .gitignore
@@ -1,13 +1,16 @@
 .DS_Store
-.idea/
+.idea
 .python-version
+.vscode
+.ruff_cache
+.pytest_cache
 
 current-workflow.txt
 package-lock.json
 
 node_modules/
-flask-app/myenv/
-flask-app/.pytest_cache
+flask_app/myenv/
+flask_app/.pytest_cache
 
 # Jupyter Notebook
 .ipynb_checkpoints
