--- flask-app/LICENSE.md
@@ -1,77 +1,0 @@
-SEED Platform™, Copyright (c) 2017, 2023 Alliance for Sustainable Energy, LLC, and other contributors.
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification, are permitted
-provided that the following conditions are met:
-
-(1) Redistributions of source code must retain the above copyright notice, this list of
-conditions and the following disclaimer.
-
-(2) Redistributions in binary form must reproduce the above copyright notice, this list of
-conditions and the following disclaimer in the documentation and/or other materials provided
-with the distribution.
-
-(3) Neither the name of the copyright holder nor the names of its contributors may be used
-to endorse or promote products derived from this software without specific prior written
-permission.
-
-(4) Other than as required in clauses (1) and (2), distributions in any form of modifications
-or other derivative works may not use the "SEED Platform" trademark, "Standard Energy
-Efficiency Data Platform", "Standard Energy Efficiency Data", "SEED", or any other confusingly
-similar designation without specific prior written permission from the U.S. Department of Energy.
-
-(5) The name of the copyright holder(s), any contributors, the United States Government, the
-United States Department of Energy, or any of their employees may not be used to endorse or
-promote products derived from this software without specific prior written permission from the
-respective party.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDER(S) AND ANY CONTRIBUTORS "AS IS" AND ANY
-EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
-MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-COPYRIGHT HOLDER(S), ANY CONTRIBUTORS, THE UNITED STATES GOVERNMENT, OR THE UNITED STATES
-DEPARTMENT OF ENERGY, NOR ANY OF THEIR EMPLOYEES, BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
-THE POSSIBILITY OF SUCH DAMAGE.
-
----
-
-This program also includes the following licenses:
-
-Copyright (c) 2014 - 2017 The Regents of the University of California, through
-Lawrence Berkeley National Laboratory (subject to receipt of any required approvals
-from the U.S. Department of Energy) and contributors. All rights reserved.
-
-1.  Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions are met:
-
-        (1) Redistributions of source code must retain the copyright notice, this
-        list of conditions and the following disclaimer.
-
-        (2) Redistributions in binary form must reproduce the copyright notice,
-        this list of conditions and the following disclaimer in the documentation
-        and/or other materials provided with the distribution.
-
-        (3) Neither the name of the University of California, Lawrence Berkeley
-        National Laboratory, U.S. Dept. of Energy nor the names of its
-        contributors may be used to endorse or promote products derived from this
-        software without specific prior written permission.
-
-        (4) Neither the names Standard Energy Efficiency Data Platform, Standard
-        Energy Efficiency Data, SEED Platform, SEED, derivatives thereof nor
-        designations containing these names, may be used to endorse or promote
-        products derived from this software without specific prior written
-        permission from the U.S. Dept. of Energy.
-
-2.  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-    AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-    IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-    ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-    (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-    LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-    ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
