+++ flask_app/.gitignore
@@ -7,7 +7,8 @@ env/
 
 dist
 
-data/
+# cached downloaded data
+/data
 .venv/
 .DS_Store
 .ipynb_checkpoints
