+++ flask_app/utils/normalize_state.py
