+++ pyproject.toml
@@ -1,12 +1,15 @@
 [tool.poetry]
-name = "cbl-web-tool"
+name = "cbl_web_tool"
 version = "0.1.0"
 description = ""
 authors = ["Devin Brady <devin.brady@nrel.gov>", "Alex Swindler <axelstudios>", "Nicholas Long <nllong>", "Roger Braun <roger.braun@nrel.gov>", "Rahman Mian <rahman.mian@nrel.gov>"]
 readme = "README.md"
 # should probably rename utils to cbl_workflow, then place
 # the utils in that folder.
-packages = [{include = "utils"}]
+packages = [
+    { include = "flask_app" },
+    { include = "flask_app/utils" }
+]
 
 [tool.poetry.dependencies]
 python = ">=3.9, <3.13"
